# api/docs/swagger.yaml
definitions:
  models.AttractionModel:
    properties:
      attraction_id:
        type: string
      attraction_name:
        type: string
      contact_number:
        type: string
      created_at:
        type: string
      description:
        type: string
      images:
        items:
          $ref: '#/definitions/models.ImageModel'
        type: array
      licence_url:
        type: string
      location:
        $ref: '#/definitions/models.LocationModel'
      owner_id:
        type: string
      rating:
        type: number
      updated_at:
        type: string
      website_url:
        type: string
    type: object
  models.CreateAttraction:
    properties:
      attraction_name:
        default: Anhor Park
        type: string
      contact_number:
        default: +(99891)-234-56-78
        type: string
      description:
        default: available for all ages
        type: string
      images:
        items:
          $ref: '#/definitions/models.CreateImage'
        type: array
      licence_url:
        default: https://creativecommons.org/licenses/by/4.0/
        type: string
      location:
        $ref: '#/definitions/models.CreateLocation'
      rating:
        default: 4.3
        type: number
      website_url:
        default: https://creativecommons.org/licenses/by/4.1/
        type: string
    type: object
...
  models.CreateImage:
    properties:
      image_url:
        default: www.photo/images/141
        type: string
    type: object
  models.CreateLocation:
    properties:
      address:
        default: 87G9+V57, Shaykhontohur Street 28, Tashkent, Toshkent Shahri
        type: string
      city:
        default: Tashkent
        type: string
      country:
        default: Uzbekistan
        type: string
      latitude:
        default: 40.7128
        type: number
      longitude:
        default: 74.006
        type: number
      state_province:
        default: Shaykhontohur
        type: string
    type: object
...
  models.DeleteResponse:
    properties:
      success:
        type: boolean
    type: object
  models.Error:
    properties:
      message:
        type: string
    type: object
...
  models.ImageModel:
    properties:
      created_at:
        type: string
      establishment_id:
        type: string
      image_id:
        type: string
      image_url:
        type: string
      updated_at:
        type: string
    type: object
  models.ListAttractionModel:
    properties:
      attractions:
        items:
          $ref: '#/definitions/models.AttractionModel'
        type: array
      overall:
        type: integer
    type: object
...
  models.LocationModel:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      establishment_id:
        type: string
      latitude:
        type: number
      location_id:
        type: string
      longitude:
        type: number
      state_province:
        type: string
      updated_at:
        type: string
    type: object
...
  models.StandartError:
    properties:
      error:
        $ref: '#/definitions/models.Error'
    type: object
...
  models.UpdateAttraction:
    properties:
      attraction_name:
        default: updated attraction name
        type: string
      contact_number:
        default: updated contact number
        type: string
      description:
        default: updated description
        type: string
      licence_url:
        default: updated licence url
        type: string
      location:
        $ref: '#/definitions/models.UpdateLocation'
      rating:
        default: 5
        type: number
      website_url:
        default: updated website url
        type: string
    type: object
...
  models.UpdateLocation:
    properties:
      address:
        default: updated address
        type: string
      city:
        default: updated city
        type: string
      country:
        default: updated country
        type: string
      latitude:
        default: 1.1
        type: number
      longitude:
        default: 1.1
        type: number
      state_province:
        default: updated state or province
        type: string
    type: object
...
info:
  contact: {}
  description: API for Touristan
  title: Welcome To Booking API
paths:
...
  /v1/attraction/create:
    post:
      consumes:
      - application/json
      description: Api for creating attraction
      parameters:
      - description: owner_id
        in: query
        name: owner_id
        required: true
        type: string
      - description: Attraction
        in: body
        name: Attraction
        required: true
        schema:
          $ref: '#/definitions/models.CreateAttraction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AttractionModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandartError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: CREATE ATTRACTION
      tags:
      - ATTRACTION
  /v1/attraction/delete:
    delete:
      consumes:
      - application/json
      description: Api for deleting attraction by attraction_id
      parameters:
      - description: attraction_id
        in: query
        name: attraction_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeleteResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandartError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: DELETE ATTRACTION BY ATTRACTION_ID
      tags:
      - ATTRACTION
  /v1/attraction/get:
    get:
      consumes:
      - application/json
      description: Api for getting attraction by attraction_id
      parameters:
      - description: attraction_id
        in: query
        name: attraction_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AttractionModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandartError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: GET ATTRACTION BY ATTRACTION_ID
      tags:
      - ATTRACTION
  /v1/attraction/list:
    get:
      consumes:
      - application/json
      description: Api for listing attractions by page and limit
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListAttractionModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandartError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: LIST ATTRACTIONS BY PAGE AND LIMIT
      tags:
      - ATTRACTION
  /v1/attraction/listbylocation:
    get:
      consumes:
      - application/json
      description: Api for listing attractions by page, limit, country, city and state_province
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      - description: country
        in: query
        name: country
        required: true
        type: string
      - description: city
        in: query
        name: city
        required: true
        type: string
      - description: state_province
        in: query
        name: state_province
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListAttractionModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandartError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: LIST ATTRACTIONS BY PAGE, LIMIT, COUNTRY, CITY AND STATE_PROVINCE
      tags:
      - ATTRACTION
  /v1/attraction/update:
    patch:
      consumes:
      - application/json
      description: Api for updating attraction by attraction_id
      parameters:
      - description: attraction_id
        in: query
        name: attraction_id
        required: true
        type: string
      - description: UpdatingAttraction
        in: body
        name: UpdatingAttraction
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAttraction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AttractionModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandartError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: UPDATE ATTRACTION
      tags:
      - ATTRACTION
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"